auth_service/pyproject.toml
[tool.poetry]
name = "auth_service"
version = "0.1.0"
description = ""
authors = ["Avex-Evo-2244 <ghbdtn2266@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
alembic = "^1.7.6"
ormar = {extras = ["mysql"], version = "^0.10.24"}
requests = "^2.27.1"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
